Separate ReactDOM host config
